fix width of variable argument in P/Invoke declarations

+++ WindowsInput/Native/NativeMethods.cs
@@ -44,7 +44,7 @@ namespace WindowsInput.Native
         /// These left- and right-distinguishing constants are only available when you call the GetKeyboardState, SetKeyboardState, GetAsyncKeyState, GetKeyState, and MapVirtualKey functions. 
         /// </remarks>
         [DllImport("user32.dll", SetLastError = true)]
-        public static extern Int16 GetAsyncKeyState(UInt16 virtualKeyCode);
+        public static extern Int16 GetAsyncKeyState(UInt32 virtualKeyCode);
 
         /// <summary>
         /// The GetKeyState function retrieves the status of the specified virtual key. The status specifies whether the key is up, down, or toggled (on, off alternating each time the key is pressed). (See: http://msdn.microsoft.com/en-us/library/ms646301(VS.85).aspx)
@@ -73,7 +73,7 @@ namespace WindowsInput.Native
         /// These left- and right-distinguishing constants are available to an application only through the GetKeyboardState, SetKeyboardState, GetAsyncKeyState, GetKeyState, and MapVirtualKey functions. 
         /// </remarks>
         [DllImport("user32.dll", SetLastError = true)]
-        public static extern Int16 GetKeyState(UInt16 virtualKeyCode);
+        public static extern Int16 GetKeyState(UInt32 virtualKeyCode);
 
         /// <summary>
         /// The SendInput function synthesizes keystrokes, mouse motions, and button clicks.
